Interface package: move MainDashboardButton slot (#27213)
Move MainDashboardButton slot from interface to appropriate
places in edit-post and edit-site packages.
