+++ docs/designers-developers/developers/slotfills/main-dashboard-button.md
@@ -1,18 +1,19 @@
 # MainDashboardButton
 
-This slot allows replacing the default main dashboard button in the post editor 
-that's used for closing the editor in fullscreen mode. In the site editor this slot
-refers to the "back to dashboard" button in the navigation sidebar.
+This slot allows replacing the default main dashboard button in the post editor and site editor. 
+It's used for returning back to main wp-admin dashboard when editor is in fullscreen mode.
 
 ## Examples
 
-Basic usage:
+### Post editor example
+
+This will override the W icon button in the header.
 
 ```js
 import { registerPlugin } from '@wordpress/plugins';
 import {
 	__experimentalMainDashboardButton as MainDashboardButton,
-} from '@wordpress/interface';
+} from '@wordpress/edit-post';
 
 const MainDashboardButtonTest = () => (
     <MainDashboardButton>
@@ -32,16 +33,14 @@ the post editor, that can be achieved in the following way:
 import { registerPlugin } from '@wordpress/plugins';
 import {
 	__experimentalFullscreenModeClose as FullscreenModeClose,
-} from '@wordpress/edit-post';
-import {
 	__experimentalMainDashboardButton as MainDashboardButton,
-} from '@wordpress/interface';
+} from '@wordpress/edit-post';
 import { close } from '@wordpress/icons';
 
 
 const MainDashboardButtonIconTest = () => (
     <MainDashboardButton>
-        <FullscreenModeClose icon={ close } />
+        <FullscreenModeClose icon={ close } href="http://wordpress.org" />
     </MainDashboardButton>
 );
 
@@ -50,13 +49,15 @@ registerPlugin( 'main-dashboard-button-icon-test', {
 } );
 ```
 
-Site editor example:
+### Site editor example
+
+In the site editor this slot refers to the "back to dashboard" button in the navigation sidebar.
 
 ```js
 import { registerPlugin } from '@wordpress/plugins';
 import {
 	__experimentalMainDashboardButton as MainDashboardButton,
-} from '@wordpress/interface';
+} from '@wordpress/edit-site';
 import {
 	__experimentalNavigationBackButton as NavigationBackButton,
 } from '@wordpress/components';
