+++ packages/edit-post/src/components/header/fullscreen-mode-close/index.js
@@ -12,7 +12,7 @@ import { __ } from '@wordpress/i18n';
 import { addQueryArgs } from '@wordpress/url';
 import { wordpress } from '@wordpress/icons';
 
-function FullscreenModeClose( { showTooltip } ) {
+function FullscreenModeClose( { showTooltip, icon, href } ) {
 	const { isActive, isRequestingSiteIcon, postType, siteIconUrl } = useSelect(
 		( select ) => {
 			const { getCurrentPostType } = select( 'core/editor' );
@@ -50,16 +50,26 @@ function FullscreenModeClose( { showTooltip } ) {
 				src={ siteIconUrl }
 			/>
 		);
-	} else if ( isRequestingSiteIcon ) {
+	}
+
+	if ( isRequestingSiteIcon ) {
 		buttonIcon = null;
+	}
+
+	// Override default icon if custom icon is provided via props.
+	if ( icon ) {
+		buttonIcon = <Icon size="36px" icon={ icon } />;
 	}
 
 	return (
 		<Button
 			className="edit-post-fullscreen-mode-close has-icon"
-			href={ addQueryArgs( 'edit.php', {
-				post_type: postType.slug,
-			} ) }
+			href={
+				href ??
+				addQueryArgs( 'edit.php', {
+					post_type: postType.slug,
+				} )
+			}
 			label={ get( postType, [ 'labels', 'view_items' ], __( 'Back' ) ) }
 			showTooltip={ showTooltip }
 		>
