+++ packages/interface/src/components/index.js
@@ -3,5 +3,4 @@ export { default as ComplementaryAreaMoreMenuItem } from './complementary-area-m
 export { default as FullscreenMode } from './fullscreen-mode';
 export { default as InterfaceSkeleton } from './interface-skeleton';
 export { default as PinnedItems } from './pinned-items';
-export { default as __experimentalMainDashboardButton } from './main-dashboard-button';
 export { default as ActionItem } from './action-item';
